refactor : 채팅 버튼 모든 페이지에 띄우기
